stake-pool: Refresh blockhash more often on update tests (#3853)
* stake-pool: Refresh blockhash more often on update tests

* Bump down max pool size, updated correctly in #3839

* Move withdraw test to another file

* Refactor withdraw tests more

* Get more blockhashes during `update_stake_pool_balance`
