+++ stake-pool/program/tests/update_validator_list_balance.rs
@@ -7,7 +7,7 @@ use {
     helpers::*,
     solana_program::{borsh::try_from_slice_unchecked, program_pack::Pack, pubkey::Pubkey},
     solana_program_test::*,
-    solana_sdk::{signature::Signer, stake::state::StakeState},
+    solana_sdk::{hash::Hash, signature::Signer, stake::state::StakeState},
     spl_stake_pool::{
         state::{StakePool, StakeStatus, ValidatorList},
         MAX_VALIDATORS_TO_UPDATE, MINIMUM_RESERVE_LAMPORTS,
@@ -20,6 +20,7 @@ async fn setup(
     num_validators: usize,
 ) -> (
     ProgramTestContext,
+    Hash,
     StakePoolAccounts,
     Vec<ValidatorStakeAccount>,
     Vec<DepositStakeAccount>,
@@ -110,12 +111,18 @@ async fn setup(
         )
         .await;
 
+    let last_blockhash = context
+        .banks_client
+        .get_new_latest_blockhash(&context.last_blockhash)
+        .await
+        .unwrap();
+
     for deposit_account in &mut deposit_accounts {
         deposit_account
             .deposit_stake(
                 &mut context.banks_client,
                 &context.payer,
-                &context.last_blockhash,
+                &last_blockhash,
                 &stake_pool_accounts,
             )
             .await;
@@ -128,7 +135,7 @@ async fn setup(
         .update_all(
             &mut context.banks_client,
             &context.payer,
-            &context.last_blockhash,
+            &last_blockhash,
             stake_accounts
                 .iter()
                 .map(|v| v.vote.pubkey())
@@ -138,8 +145,15 @@ async fn setup(
         )
         .await;
 
+    let last_blockhash = context
+        .banks_client
+        .get_new_latest_blockhash(&context.last_blockhash)
+        .await
+        .unwrap();
+
     (
         context,
+        last_blockhash,
         stake_pool_accounts,
         stake_accounts,
         deposit_accounts,
@@ -154,6 +168,7 @@ async fn success_with_normal() {
     let num_validators = 5;
     let (
         mut context,
+        last_blockhash,
         stake_pool_accounts,
         stake_accounts,
         _,
@@ -193,11 +208,17 @@ async fn success_with_normal() {
     slot += slots_per_epoch;
     context.warp_to_slot(slot).unwrap();
 
+    let last_blockhash = context
+        .banks_client
+        .get_new_latest_blockhash(&last_blockhash)
+        .await
+        .unwrap();
+
     stake_pool_accounts
         .update_all(
             &mut context.banks_client,
             &context.payer,
-            &context.last_blockhash,
+            &last_blockhash,
             stake_accounts
                 .iter()
                 .map(|v| v.vote.pubkey())
@@ -225,8 +246,16 @@ async fn success_with_normal() {
 
 #[tokio::test]
 async fn merge_into_reserve() {
-    let (mut context, stake_pool_accounts, stake_accounts, _, lamports, _, mut slot) =
-        setup(MAX_VALIDATORS_TO_UPDATE).await;
+    let (
+        mut context,
+        last_blockhash,
+        stake_pool_accounts,
+        stake_accounts,
+        _,
+        lamports,
+        _,
+        mut slot,
+    ) = setup(MAX_VALIDATORS_TO_UPDATE).await;
 
     let pre_lamports = get_validator_list_sum(
         &mut context.banks_client,
@@ -249,7 +278,7 @@ async fn merge_into_reserve() {
             .decrease_validator_stake(
                 &mut context.banks_client,
                 &context.payer,
-                &context.last_blockhash,
+                &last_blockhash,
                 &stake_account.stake_account,
                 &stake_account.transient_stake_account,
                 lamports,
@@ -264,7 +293,7 @@ async fn merge_into_reserve() {
         .update_all(
             &mut context.banks_client,
             &context.payer,
-            &context.last_blockhash,
+            &last_blockhash,
             stake_accounts
                 .iter()
                 .map(|v| v.vote.pubkey())
@@ -295,11 +324,16 @@ async fn merge_into_reserve() {
     slot += slots_per_epoch;
     context.warp_to_slot(slot).unwrap();
 
+    let last_blockhash = context
+        .banks_client
+        .get_new_latest_blockhash(&last_blockhash)
+        .await
+        .unwrap();
     stake_pool_accounts
         .update_all(
             &mut context.banks_client,
             &context.payer,
-            &context.last_blockhash,
+            &last_blockhash,
             stake_accounts
                 .iter()
                 .map(|v| v.vote.pubkey())
@@ -336,8 +370,16 @@ async fn merge_into_reserve() {
 
 #[tokio::test]
 async fn merge_into_validator_stake() {
-    let (mut context, stake_pool_accounts, stake_accounts, _, lamports, reserve_lamports, mut slot) =
-        setup(MAX_VALIDATORS_TO_UPDATE).await;
+    let (
+        mut context,
+        last_blockhash,
+        stake_pool_accounts,
+        stake_accounts,
+        _,
+        lamports,
+        reserve_lamports,
+        mut slot,
+    ) = setup(MAX_VALIDATORS_TO_UPDATE).await;
 
     let rent = context.banks_client.get_rent().await.unwrap();
     let pre_lamports = get_validator_list_sum(
@@ -352,7 +394,7 @@ async fn merge_into_validator_stake() {
     let current_minimum_delegation = stake_pool_get_minimum_delegation(
         &mut context.banks_client,
         &context.payer,
-        &context.last_blockhash,
+        &last_blockhash,
     )
     .await;
     let available_lamports =
@@ -363,7 +405,7 @@ async fn merge_into_validator_stake() {
             .increase_validator_stake(
                 &mut context.banks_client,
                 &context.payer,
-                &context.last_blockhash,
+                &last_blockhash,
                 &stake_account.transient_stake_account,
                 &stake_account.stake_account,
                 &stake_account.vote.pubkey(),
@@ -382,7 +424,7 @@ async fn merge_into_validator_stake() {
         .update_all(
             &mut context.banks_client,
             &context.payer,
-            &context.last_blockhash,
+            &last_blockhash,
             stake_accounts
                 .iter()
                 .map(|v| v.vote.pubkey())
@@ -413,11 +455,16 @@ async fn merge_into_validator_stake() {
     slot += slots_per_epoch;
     context.warp_to_slot(slot).unwrap();
 
+    let last_blockhash = context
+        .banks_client
+        .get_new_latest_blockhash(&last_blockhash)
+        .await
+        .unwrap();
     let error = stake_pool_accounts
         .update_all(
             &mut context.banks_client,
             &context.payer,
-            &context.last_blockhash,
+            &last_blockhash,
             stake_accounts
                 .iter()
                 .map(|v| v.vote.pubkey())
@@ -475,15 +522,23 @@ async fn merge_into_validator_stake() {
 
 #[tokio::test]
 async fn merge_transient_stake_after_remove() {
-    let (mut context, stake_pool_accounts, stake_accounts, _, lamports, reserve_lamports, mut slot) =
-        setup(1).await;
+    let (
+        mut context,
+        last_blockhash,
+        stake_pool_accounts,
+        stake_accounts,
+        _,
+        lamports,
+        reserve_lamports,
+        mut slot,
+    ) = setup(1).await;
 
     let rent = context.banks_client.get_rent().await.unwrap();
     let stake_rent = rent.minimum_balance(std::mem::size_of::<StakeState>());
     let current_minimum_delegation = stake_pool_get_minimum_delegation(
         &mut context.banks_client,
         &context.payer,
-        &context.last_blockhash,
+        &last_blockhash,
     )
     .await;
     let deactivated_lamports = lamports;
@@ -493,7 +548,7 @@ async fn merge_transient_stake_after_remove() {
             .decrease_validator_stake(
                 &mut context.banks_client,
                 &context.payer,
-                &context.last_blockhash,
+                &last_blockhash,
                 &stake_account.stake_account,
                 &stake_account.transient_stake_account,
                 deactivated_lamports,
@@ -505,7 +560,7 @@ async fn merge_transient_stake_after_remove() {
             .remove_validator_from_pool(
                 &mut context.banks_client,
                 &context.payer,
-                &context.last_blockhash,
+                &last_blockhash,
                 &stake_account.stake_account,
                 &stake_account.transient_stake_account,
             )
@@ -523,7 +578,7 @@ async fn merge_transient_stake_after_remove() {
         .update_all(
             &mut context.banks_client,
             &context.payer,
-            &context.last_blockhash,
+            &last_blockhash,
             stake_accounts
                 .iter()
                 .map(|v| v.vote.pubkey())
@@ -560,7 +615,7 @@ async fn merge_transient_stake_after_remove() {
         .update_validator_list_balance(
             &mut context.banks_client,
             &context.payer,
-            &context.last_blockhash,
+            &last_blockhash,
             stake_accounts
                 .iter()
                 .map(|v| v.vote.pubkey())
@@ -613,11 +668,7 @@ async fn merge_transient_stake_after_remove() {
 
     // Update stake pool balance and cleanup, should be gone
     let error = stake_pool_accounts
-        .update_stake_pool_balance(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-        )
+        .update_stake_pool_balance(&mut context.banks_client, &context.payer, &last_blockhash)
         .await;
     assert!(error.is_none());
 
@@ -625,7 +676,7 @@ async fn merge_transient_stake_after_remove() {
         .cleanup_removed_validator_entries(
             &mut context.banks_client,
             &context.payer,
-            &context.last_blockhash,
+            &last_blockhash,
         )
         .await;
     assert!(error.is_none());
@@ -643,8 +694,16 @@ async fn merge_transient_stake_after_remove() {
 #[tokio::test]
 async fn success_with_burned_tokens() {
     let num_validators = 5;
-    let (mut context, stake_pool_accounts, stake_accounts, deposit_accounts, _, _, mut slot) =
-        setup(num_validators).await;
+    let (
+        mut context,
+        last_blockhash,
+        stake_pool_accounts,
+        stake_accounts,
+        deposit_accounts,
+        _,
+        _,
+        mut slot,
+    ) = setup(num_validators).await;
 
     let mint_info = get_account(
         &mut context.banks_client,
@@ -664,7 +723,7 @@ async fn success_with_burned_tokens() {
     burn_tokens(
         &mut context.banks_client,
         &context.payer,
-        &context.last_blockhash,
+        &last_blockhash,
         &stake_pool_accounts.token_program_id,
         &stake_pool_accounts.pool_mint.pubkey(),
         &deposit_accounts[0].pool_account.pubkey(),
@@ -690,7 +749,7 @@ async fn success_with_burned_tokens() {
         .update_all(
             &mut context.banks_client,
             &context.payer,
-            &context.last_blockhash,
+            &last_blockhash,
             stake_accounts
                 .iter()
                 .map(|v| v.vote.pubkey())
