+++ stake-pool/program/tests/withdraw.rs
@@ -18,97 +18,10 @@ use {
         transaction::{Transaction, TransactionError},
         transport::TransportError,
     },
-    spl_stake_pool::{error::StakePoolError, id, instruction, state, MINIMUM_RESERVE_LAMPORTS},
+    spl_stake_pool::{error::StakePoolError, id, instruction, state},
     spl_token::error::TokenError,
     test_case::test_case,
 };
-
-async fn setup(
-    token_program_id: Pubkey,
-) -> (
-    ProgramTestContext,
-    StakePoolAccounts,
-    ValidatorStakeAccount,
-    DepositStakeAccount,
-    Keypair,
-    Keypair,
-    u64,
-) {
-    let mut context = program_test().start_with_context().await;
-    let stake_pool_accounts = StakePoolAccounts::new_with_token_program(token_program_id);
-    stake_pool_accounts
-        .initialize_stake_pool(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            MINIMUM_RESERVE_LAMPORTS,
-        )
-        .await
-        .unwrap();
-
-    let validator_stake_account = simple_add_validator_to_pool(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &stake_pool_accounts,
-        None,
-    )
-    .await;
-
-    let current_minimum_delegation = stake_pool_get_minimum_delegation(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-    )
-    .await;
-
-    let deposit_info = simple_deposit_stake(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &stake_pool_accounts,
-        &validator_stake_account,
-        current_minimum_delegation * 3,
-    )
-    .await
-    .unwrap();
-
-    let tokens_to_withdraw = deposit_info.pool_tokens;
-
-    // Delegate tokens for withdrawing
-    let user_transfer_authority = Keypair::new();
-    delegate_tokens(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &stake_pool_accounts.token_program_id,
-        &deposit_info.pool_account.pubkey(),
-        &deposit_info.authority,
-        &user_transfer_authority.pubkey(),
-        tokens_to_withdraw,
-    )
-    .await;
-
-    // Create stake account to withdraw to
-    let user_stake_recipient = Keypair::new();
-    create_blank_stake_account(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &user_stake_recipient,
-    )
-    .await;
-
-    (
-        context,
-        stake_pool_accounts,
-        validator_stake_account,
-        deposit_info,
-        user_transfer_authority,
-        user_stake_recipient,
-        tokens_to_withdraw,
-    )
-}
 
 #[test_case(spl_token::id(); "token")]
 #[test_case(spl_token_2022::id(); "token-2022")]
@@ -136,7 +49,7 @@ async fn _success(token_program_id: Pubkey, test_type: SuccessTestType) {
         user_transfer_authority,
         user_stake_recipient,
         tokens_to_withdraw,
-    ) = setup(token_program_id).await;
+    ) = setup_for_withdraw(token_program_id).await;
 
     // Save stake pool state before withdrawal
     let stake_pool_before = get_account(
@@ -355,7 +268,7 @@ async fn fail_with_wrong_stake_program() {
         user_transfer_authority,
         user_stake_recipient,
         tokens_to_burn,
-    ) = setup(spl_token::id()).await;
+    ) = setup_for_withdraw(spl_token::id()).await;
 
     let new_authority = Pubkey::new_unique();
     let wrong_stake_program = Pubkey::new_unique();
@@ -416,7 +329,7 @@ async fn fail_with_wrong_withdraw_authority() {
         user_transfer_authority,
         user_stake_recipient,
         tokens_to_burn,
-    ) = setup(spl_token::id()).await;
+    ) = setup_for_withdraw(spl_token::id()).await;
 
     let new_authority = Pubkey::new_unique();
     stake_pool_accounts.withdraw_authority = Keypair::new().pubkey();
@@ -458,7 +371,7 @@ async fn fail_with_wrong_token_program_id() {
         user_transfer_authority,
         user_stake_recipient,
         tokens_to_burn,
-    ) = setup(spl_token::id()).await;
+    ) = setup_for_withdraw(spl_token::id()).await;
 
     let new_authority = Pubkey::new_unique();
     let wrong_token_program = Keypair::new();
@@ -510,7 +423,7 @@ async fn fail_with_wrong_validator_list() {
         user_transfer_authority,
         user_stake_recipient,
         tokens_to_burn,
-    ) = setup(spl_token::id()).await;
+    ) = setup_for_withdraw(spl_token::id()).await;
 
     let new_authority = Pubkey::new_unique();
     stake_pool_accounts.validator_list = Keypair::new();
@@ -554,7 +467,7 @@ async fn fail_with_unknown_validator() {
         user_transfer_authority,
         user_stake_recipient,
         tokens_to_withdraw,
-    ) = setup(spl_token::id()).await;
+    ) = setup_for_withdraw(spl_token::id()).await;
 
     let unknown_stake = create_unknown_validator_stake(
         &mut context.banks_client,
@@ -600,7 +513,7 @@ async fn fail_double_withdraw_to_the_same_account() {
         user_transfer_authority,
         user_stake_recipient,
         tokens_to_burn,
-    ) = setup(spl_token::id()).await;
+    ) = setup_for_withdraw(spl_token::id()).await;
 
     let new_authority = Pubkey::new_unique();
     let error = stake_pool_accounts
@@ -666,7 +579,7 @@ async fn fail_without_token_approval() {
         user_transfer_authority,
         user_stake_recipient,
         tokens_to_burn,
-    ) = setup(spl_token::id()).await;
+    ) = setup_for_withdraw(spl_token::id()).await;
 
     revoke_tokens(
         &mut context.banks_client,
@@ -718,7 +631,7 @@ async fn fail_with_not_enough_tokens() {
         user_transfer_authority,
         user_stake_recipient,
         tokens_to_burn,
-    ) = setup(spl_token::id()).await;
+    ) = setup_for_withdraw(spl_token::id()).await;
 
     // Empty validator stake account
     let empty_stake_account = simple_add_validator_to_pool(
